backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlmodel
python-dotenv
alembic

# Database
psycopg2-binary

# Auth & Security
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt
email-validator

# AI & Tools
openai
requests
httpx